# perf(nuxt): use static import of `@unhead/ssr`