<!---
☝️ PR title should follow conventional commits (https://conventionalcommits.org)

Please carefully read the contribution docs before creating a pull request
 👉 https://v3.nuxtjs.org/community/contribution
-->

### 🔗 Linked issue

relates to #9825 

### ❓ Type of change

<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply. -->

- [ ] 📖 Documentation (updates to the documentation or readme)
- [ ] 🐞 Bug fix (a non-breaking change that fixes an issue)
- [X] 👌 Enhancement (improving an existing functionality like performance)
- [ ] ✨ New feature (a non-breaking change that adds functionality)
- [ ] ⚠️ Breaking change (fix or feature that would cause existing functionality to change)

### 📚 Description

Improve SSR performance by using static import :
* Before:
```
> wrk -t 1 -d 10 -c 5 http://localhost:3000/
Running 10s test @ http://localhost:3000/
  1 threads and 5 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.53ms    1.97ms  20.17ms   78.81%
    Req/Sec     0.91k    41.85     0.97k    71.00%
  9045 requests in 10.02s, 0.91GB read
Requests/sec:    903.14
Transfer/sec:     92.82MB
```
* After:
```
> wrk -t 1 -d 10 -c 5 http://localhost:3000/
Running 10s test @ http://localhost:3000/
  1 threads and 5 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.40ms    1.75ms  18.50ms   82.62%
    Req/Sec     1.15k    63.17     1.26k    70.00%
  11447 requests in 10.01s, 1.15GB read
Requests/sec:   1143.10
Transfer/sec:    117.48MB
```

### 📝 Checklist

<!-- Put an `x` in all the boxes that apply. -->
<!-- If your change requires a documentation PR, please link it appropriately -->
<!-- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [X] I have linked an issue or discussion.
- [ ] I have updated the documentation accordingly.