@manniL 2022-12-22
As mentioned in https://github.com/nuxt/framework/issues/9825#issuecomment-1363176179, this might have unintended impact on the client bundle.