@harlan-zw 2022-12-23
As long as it's getting tree shaken out in the client build, which I'm guessing it will, then this PR looks good.