@pi0 2022-12-23
Double checking [dist](https://unpkg.com/browse/@unhead/ssr@1.0.14/dist/index.mjs) it has also no side effects so should be fine this way at least with current code.